server/requirements.txt
pipecat-ai[webrtc,silero,groq,openai,cartesia,runner]>=0.0.77
python-dotenv
loguru